Implement updating the CFG savegame after setting EULAs
After writing EULAs to the config info block, call the
`CFG_UpdateConfigSavegame()` function (which was formerly named
`CFG_UpdateConfigNANDSavegame()` in older versions of libctru).

This is a code change, mimicking the behaviour of the
EULASetter app that DSES is based on.

This also tweaks some of the messages printed to the screen.

Fixes an issue in which EULAs weren't guaranteed to persist
after the system reboots or is powered off.